+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+- Fix Visa/MasterCard payment error
+- Added referencePrefix (That let you track the type of the transaction)
 ## 0.0.27
 
 - Update dependencies